+++ src/app/dataset-view/dataset.service.spec.ts
@@ -205,7 +205,7 @@ describe("AppDatasetService", () => {
 
         const subscriptionErrorChangesSpy = spyOn(appDatasetSubsService, "observeSqlErrorOccurred");
 
-        service.requestDatasetDataSqlRun(query, limit).subscribe();
+        service.requestDatasetDataSqlRun({ query, limit }).subscribe();
 
         expect(subscriptionDataChanges$.closed).toBeTrue();
         expect(subscriptionErrorChangesSpy).toHaveBeenCalledWith({ error: "" });
@@ -220,14 +220,14 @@ describe("AppDatasetService", () => {
             /* Intentionally blank */
         });
 
-        const subscriptionErrorChanges$ = appDatasetSubsService.onDatasetDataSqlErrorOccured
+        const subscriptionErrorChanges$ = appDatasetSubsService.onDatasetDataSqlErrorOccurred
             .pipe(first())
             .subscribe((update: DataSqlErrorUpdate) => {
                 const errorResult = mockDatasetDataSqlRunInvalidSqlResponse.data.query as DataQueryResultError;
                 expect(update.error).toEqual(errorResult.errorMessage);
             });
 
-        service.requestDatasetDataSqlRun(query, limit).subscribe();
+        service.requestDatasetDataSqlRun({ query, limit }).subscribe();
 
         expect(subscriptionDataChanges$.closed).toBeFalse();
         expect(subscriptionErrorChanges$.closed).toBeTrue();
@@ -239,10 +239,10 @@ describe("AppDatasetService", () => {
         spyOn(datasetApi, "getDatasetDataSqlRun").and.returnValue(of(mockDatasetDataSqlRunInternalErrorResponse));
 
         appDatasetSubsService.onDatasetDataChanges.subscribe(() => fail("Unexpected data update"));
-        appDatasetSubsService.onDatasetDataSqlErrorOccured.subscribe(() => fail("Unexpected SQL error update"));
+        appDatasetSubsService.onDatasetDataSqlErrorOccurred.subscribe(() => fail("Unexpected SQL error update"));
 
         const subscription$ = service
-            .requestDatasetDataSqlRun(query, limit)
+            .requestDatasetDataSqlRun({ query, limit })
             .pipe(first())
             .subscribe(
                 () => {
@@ -262,10 +262,10 @@ describe("AppDatasetService", () => {
         spyOn(datasetApi, "getDatasetDataSqlRun").and.returnValue(throwError(new SqlExecutionError()));
 
         appDatasetSubsService.onDatasetDataChanges.subscribe(() => fail("Unexpected data update"));
-        appDatasetSubsService.onDatasetDataSqlErrorOccured.subscribe(() => fail("Unexpected SQL error update"));
+        appDatasetSubsService.onDatasetDataSqlErrorOccurred.subscribe(() => fail("Unexpected SQL error update"));
 
         const subscription$ = service
-            .requestDatasetDataSqlRun(query, limit)
+            .requestDatasetDataSqlRun({ query, limit })
             .pipe(first())
             .subscribe(
                 () => {