+++ src/app/interface/dataset.interface.ts
@@ -26,3 +26,9 @@ export interface DatasetsAccountResponse {
     pageInfo: PageBasedInfo;
     datasetTotalCount: number;
 }
+
+export interface DatasetRequestBySql {
+    query: string;
+    limit?: number;
+    skip?: number;
+}